Fixed pkg name, fixed Get/Post, added user agent.
First of all, I changed the package name from "go-mwclient" to
"mwclient". The compiler didn't like the previous name and this name is
better anyway.

API.Get()/Post() now set a user agent (a link to go-mwclient's GitHub
repository). This turned out not to be easy using the standard
http.Client.Get/Post methods, so I had to do a lot of work to make them
work properly.