Added ErrorCheck func for checking for API errors.
ErrorCheck can be used as a wrapper for Wiki.Call and will return a bool
as its last return value (in addition to passing through the
simplejson.Json and error parameters).

If the bool (called apiok internally in ErrorCheck) is false, an API
error or warning has been found. If true, no errors or warnings have
been found.