Fix maxlag support
The maxlag code in Client.call() didn't actually do anything (I don't
know what I was thinking when I added that code). However, now it does
properly add the maxlag parameter to the URL. To ensure that the maxlag
parameter is the last in the URL, the URL is now encoded using
URLEncode() from the new file helpers.go. URLEncode() is a modified
version of Values.Encode() from net/url that adds maxlag as the last
parameter when applicable.