Add Page type, fix GetPage to return a *Page
Instead of making GetPage return another value, I've made a Page struct
containing two fields: Body and Timestamp, which GetPage now, if the
request is successful, returns.

The Timestamp contains the timestamp of the latest revision to the page,
and it should be passed as a parameter in any edits to detect edit
conflicts.

The diff looks dramatic because I moved GetToken to the bottom of the
file.