Overhaul API error handling
Add ExtractAPIErrors function which extracts API errors and warnings, if
any, from a *simplejson.Json object and returns them stored in a
multierror.Errors object from the github.com/joeshaw/multierror package.

ExtractAPIErrors will probably complain if the API response contains a
malformed warning. Although this is unlikely to happen, I should
probably fix it.

ErrorCheck is now deprecated and will be removed in a later commit.
GetCheck and PostCheck are removed entirely in this commit.

TODO: Testing... this code really needs to be tested more thoroughly.