Clean up exported interface, comments
Make these functions/types unexported:
- CAPTCHAError (now captchaError)
- MaxLagError (now MaxLagError)
- URLEncode (now urlEncode)

Remove descriptions of how the call is performed in the doc comments
for Client.Get and Client.Post as this information is mostly irrelevant
to users (and if it isn't, they're going to look at the source anyway).

Also make a minor, unimportant change to a comment in Client.call.