Clean up Client constructors
New now takes the same parameters as NewDefault (which has been
removed). New takes an API URL and a partial user agent and
returns a *Client and an error. Maxlag settings must be configured
after construction by setting the fields in Client.Maxlag.

New returns an error if the inURL argument is invalid or if the
userAgent argument is empty.

Update the test for this change.