Replace Client.GetPage
Client.GetPage required the pageid of the desired page, which meant that
usually one would have to first make one API call to get the pageid from
the name and then call GetPage. With the new getPage method, one can
make only one call to the API to get the desired page. getPage is
wrapped by the GetPageByName and GetPageByID methods which take a page
name and a pageid, respectively.