Support requests for raw JSON API responses
Change Client.call to return the raw JSON API response (as a []byte)
instead of decoding it into a simplejson.Json object. If such an object
is desired, Client.callJSON, which wraps Client.call and decodes the
JSON into a simplejson.Json object, can be used. Client.callJSON is
equivalent to calling Client.call before this commit.

Add GetRaw and PostRaw methods for Client, which return the raw JSON
response as a []byte. These methods do not extract API errors/warnings
like the Get and Post methods do.

Change/add a couple of comments.