Use indexpageids param in getPage method
This makes the API return the ids of page objects in the JSON response
in a JSON array to allow for simpler iteration (or in this case
extracting a single pageid). This change is similiar to what was done to
the GetPageID method in an earlier commit.

Also gave getPage and its wrapper methods named return values to make it
more clear in what order the content and timestamp is returned.