Minor changes to keep up to date with MediaWiki
The MediaWiki API is switching to a new query continuation system, and
because this change is backwards incompatible and will eventually be the
default, any time you make a query it returns a warning about how the
current syntax is deprecated. Adding an empty "rawcontinue" parameter to
query requests suppresses this warning and keeps using the old system.
I'll probably switch to the new system if I need to handle continuations.

Also, a minor cosmetic change: renamed a local variable to reflect the
edit token being renamed 'csrf' in the API.