Add consts for API token names
These consts (mwclient.CSRFToken, etc.) are meant to be used with the
GetToken method as the tokenName argument instead of typing the token name
manually each time GetToken is called.

This is to provide type-safety to the tokenName parameter to avoid typos
when writing token names manually. Furthermore, it also facilitates
autocompletion and looking up token names without going to the MW API
docs.