Remove GetPageID(), add GetPageIDs()
Client.GetPageIDs() takes a list of page names as argument and returns a
map[string]string where the key is the page name and the value is the page
ID.

This promotes courteous API usage as one would make a single call for all
the page IDs one wanted instead of one call for each.

However, most (if not all) API calls can accept both IDs and names, so
this method is pretty much useless. As I recall, it is a relic from an
earlier design of go-mwclient that I decided against.

Also, there's an annoying amount of type checking in the method because it
uses go-simplejson instead of decoding it into a struct with the stdlib
json package.

I also added an error variable ErrNoArgs, moved some stuff around in
errors.go, and implemented a golint suggestion.