Refactor API error handling
The old system of reporting API errors/warnings was pretty horrible, so
I've refactored the central piece of the system: extractAPIErrors().

It now only returns an error, it does not pass through the simplejson
object. Instead of returning a multierror, it will now simply return an
APIError if it detects one (I don't think more than one error can be
returned in an API response). If it doesn't, but it detects one or more
warnings, it will return an error of type APIWarnings, which is a
"subtype" (I don't know what else to call it) of slice, so it can be
ranged over.

The old APIWarning type has been removed. As has the vendored multierror
package and the IsAPIErr helper function.

The test for extractAPIErrors has been updated to fit the refactored
function.