Fix duplicate cookie jar
For some reason Client had both a cookiejar in Client itself and in the
http client it contains.

Remove Client.cjar and change dump/load cookie methods to use the cookie
jar in Client.httpc.Jar.

In New(), check the error returned by cookiejar.New(). It looks like it
can't actually fail, but it could change later, so just to be safe.