Update Login() for changed MW API
Getting login tokens via action=login was recently deprecated in the
MediaWiki API. Change Login() to use GetToken() for retrieving login
tokens.

Change GetToken() so that it will not use the token cache for login
tokens. It doesn't make sense to store it because you won't reuse it.

The diff for core_test.go looks crazy, but it's actually not that complex
a change. Remove the TestLogin test because it didn't work following the
changes to Login(). Add TestLoginToken test, that tests that the client
properly asks for a login token and then uses that token in a login
request.