Clarify error checking in getPage
Invert the error checking, so that the error path is indented rather
than the happy path.

These are the expected error cases:

 - If getPages returns a nil map and an error, return the error.

 - If getPages returns a non-nil map and an error, then the error should
 be an API warning. Return the warning along with the page content and
 timestamp if there are no other errors.

 - If getPages returns a non-nil map and a nil error and the
 BriefRevision in the map contains an error, then return the
 BriefRevision error.

getPages should never return a nil map and a nil error, so don't bother
handling that case explicitly. getPage will panic when it tries to use
the nil map. Likewise, if getPages returns a non-nil map but it doesn't
have an entry for the page id/name, getPage will also panic.